While we're here, let's clean up DisplayStatus as well
